refactor(Day1): merge into one file with Container class
